Separate dashboard signal calculations from web handling
